requirements.txt
openai>=1.0.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0